Fixed issues with NeuroForging UI Element
Fixed issues with NeuroForging UI Element not displaying when attempting to NF an item. Also changed Default Directory to Custom. This will make it so you do not have to override your original Default Folder to install the mod, Just need to copy paste the Custom Folder alongside the Default Folder in the Star Sonata 2 Content Directory.